requirements.txt
ansible-base==2.10.8
apache-libcloud==3.2.0
APScheduler==3.9.1
argcomplete==1.8.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
boto3==1.24.35
botocore==1.27.35
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
cloud-init==22.2
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
dnspython==2.1.0
ec2-hibinit-agent==1.0.0
Flask==2.1.3
Flask-Cors==3.0.10
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.0.3
jmespath==0.10.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
MarkupSafe==2.0.1
more-itertools==8.10.0
netaddr==0.8.0
netifaces==0.11.0
ntlm-auth==1.4.0
oauthlib==3.2.0
packaging==21.3
pexpect==4.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycryptodomex==3.11.0
PyGObject==3.42.0
PyHamcrest==2.0.2
pyinotify==0.9.6
PyJWT==2.4.0
pykerberos==1.1.14
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.3.0+ubuntu2
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywinrm==0.3.0
PyYAML==5.4.1
requests==2.25.1
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
s3transfer==0.6.0
SecretStorage==3.3.1
selinux==3.3
service-identity==18.1.0
simplejson==3.17.6
six==1.16.0
sos==4.3
ssh-import-id==5.11
systemd-python==234
Twisted==22.1.0
tzdata==2022.1
tzlocal==4.2
ubuntu-advantage-tools==27.9
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
Werkzeug==2.1.2
xmltodict==0.12.0
zipp==1.0.0
zope.interface==5.4.0
